pyproject.toml
# =============================================================================
# Minute - Project Configuration (UV Compatible - PEP 621)
# =============================================================================

[project]
name = "minute"
version = "2.0.0"
description = "Meeting transcription and minute generation application"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "i.AI", email = "i-dot-ai-enquiries@cabinetoffice.gov.uk"}
]
requires-python = ">=3.12,<3.13"

# Core dependencies (shared between backend and worker)
dependencies = [
    "cryptography>=44.0.2",
    "pydantic-settings>=2.5.2",
    "alembic>=1.13.3",
    "sqlmodel>=0.0.27",
    "tenacity>=9.1.2",
    "aioboto3>=14.3.0",
    "breame>=0.1.2",
    "mistune>=3.1.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.9",
    "boto3>=1.35.3",
    "fastapi>=0.112.0",
    "sentry-sdk[fastapi]>=2.18.0",
    "sqlalchemy[asyncio]>=2.0.42",
    "asyncpg>=0.30.0",
    "markdownify>=1.2.0",
    "python-multipart>=0.0.20",
    "i-dot-ai-utilities>=0.4.4",
    "azure-servicebus>=7.14.2",
]

[project.optional-dependencies]
# Backend-specific dependencies
backend = [
    "pyjwt>=2.9.0",
    "posthog>=6.6.1",
    "uvicorn>=0.31.0",
    "apscheduler>=3.11.0",
]

# Worker-specific dependencies
worker = [
    "google-genai>=1.10.0",
    "openai>=1.42.0",
    "ffmpeg-python>=0.2.0",
    "ray[default]>=2.53.0",
    "azure-storage-blob>=12.26.0",
    "azure-identity>=1.23.1",
]

# Local model dependencies (for running Minute entirely locally)
local = [
    "ollama>=0.4.0",
    "faster-whisper>=1.0.0",
    "ctranslate2>=4.0.0",
]

# Development dependencies
dev = [
    "bumpversion>=0.6.0",
    "pre-commit>=3.8.0",
    "pytest>=8.3.2",
    "pytest-env>=1.1.1",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "pytest-dotenv>=0.5.2",
    "pytest-asyncio>=0.23.8",
    "ruff==0.8.2",
    "bandit>=1.7.9",
    "detect-secrets>=1.5.0",
    "playwright>=1.45",
    "pytest-playwright>=0.5",
    "diagrams>=0.23.4",
    "ipykernel>=6.29.5",
    "beautifulsoup4>=4.13.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["worker", "backend", "common"]

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.ruff]
line-length = 120
target-version = 'py312'
fix = true

[tool.ruff.format]
exclude = ["data/*"]

[tool.ruff.lint]
select = [
    "A", "B", "C", "E", "F", "G", "I", "N", "Q", "S", "T", "W",
    "ARG", "ASYNC", "BLE", "COM", "DJ", "DTZ", "EM", "EXE", "FURB",
    "ICN", "INT", "ISC", "LOG", "NPY", "PD", "PGH", "PIE", "PL",
    "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "SLOT",
    "TCH", "TD", "TID", "UP", "YTT"
]
exclude = ["out/*"]
ignore = ["COM812", "DJ001", "RET505", "RET508", "PLR0913", "S318", "ISC001"]

[tool.ruff.lint.isort]
known-first-party = ["backend", "frontend"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "PLR0913", "PLR0915", "PLR2004", "TD003", "S311"]
"*/tests/*" = ["S101", "S106", "PLR0913", "PLR0915", "PLR2004", "TD003"]
"*/tests_playwright/*" = ["S101", "S106", "PLR0913", "PLR2004", "TD003"]

[tool.cruft]
skip = [
    ".env.*",
    "data",
]

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra -q"
testpaths = ["tests"]
